pyproject.toml
[tool.poetry]
name = "https-ecdsa-sigdump"
version = "0.1.0"
description = ""
authors = ["Nils Amiet <nils.amiet@kudelskisecurity.com>"]

[tool.poetry.dependencies]
python = "^3.10"
scapy = "^2.4.5"
cryptography = "^37.0.4"
pycryptodome = "^3.15.0"
ecdsa = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
